Add types to ingredients
